Move fetch of third-party API to server side
