+++ fashion-messaging-app-api/server.js
@@ -1,3 +1,4 @@
+import { createClient } from "pexels";
 import express from 'express';
 import session from 'express-session';
 import cors from 'cors';
@@ -16,12 +17,64 @@ const __filename = fileURLToPath(import.meta.url)
 const __dirname = dirname(__filename)
 const app = express();
 
+
+const API_KEY = "06ZcwjgbmJBM8T2TxLUZ5iwdXXGxiAgz0Z018b7QPKwR1ExipkFjaAuw";
+const clientAPI = createClient(API_KEY);
+const defaultQuery = "Fashion";
+const PHOTOS = 1;
+
+
 app.use(cors({
   origin: 'http://localhost:3001',
   credentials: true
 }));
 app.use(express.json()); 
 app.use(morgan('combined')); 
+
+
+app.get('/photos', async (req, res) => {
+  const query = req.query.search || defaultQuery;
+  
+  try {
+    const response = await clientAPI.photos.search({ query, per_page: PHOTOS });
+   
+    var myHeaders = new Headers();
+    myHeaders.append("x-api-key", "4d552e9f30522b1dec7c712f83c67c235be86e25ec14b4ba3493383ec7b81d3f");                               
+
+    let fetchPromises = await Promise.all(response.photos.map(async photo => { 
+      var formdata = new FormData();
+      formdata.append("image_url", photo.src.original);
+    
+
+      var requestOptions = {
+        method: 'POST',
+        headers: myHeaders,
+        body: formdata,
+      };
+
+      try {
+        const fetchResponse = await fetch('https://cloudapi.lykdat.com/v1/detection/items', requestOptions);
+        const result = await fetchResponse.json();     
+
+        if (result.data.detected_items.length > 0) {
+          return photo;     
+        }
+       
+      } catch (error) {
+        console.error('error', error);
+      }
+     
+    }));
+
+    const fashionPhotos = fetchPromises.filter(photo => photo !== undefined)
+    res.json(fashionPhotos);
+
+  } catch (error) {
+    console.error("Error fetching photos:", error);
+    res.status(500).json({message: error.message});
+  }
+});
+
 
 const storage = multer.diskStorage({
   destination: function(req,file,cb) {
@@ -113,4 +166,3 @@ sequelize.sync({ alter: true })
   .catch(error => {
     console.error('Unable to connect to the database:', error);
   });
-
