+++ fashion-messaging-app-ui/src/App.jsx
@@ -1,7 +1,6 @@
 
 import "./App.css";
 import { useEffect, useState } from "react";
-import { createClient } from "pexels";
 import { UserContext } from './UserContext';
 import Main from './components/Main/Main'
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
@@ -9,10 +8,6 @@ import LoginForm from './components/LoginForm/LoginForm';
 import SignupForm from './components/SignupForm/SignupForm';
 import Homepage from './components/Homepage/Homepage';
 
-const API_KEY = "06ZcwjgbmJBM8T2TxLUZ5iwdXXGxiAgz0Z018b7QPKwR1ExipkFjaAuw";
-const clientAPI = createClient(API_KEY);
-const defaultQuery = "Fashion";
-const PHOTOS = 1;
 
 export default function App() {
   const [loading, setLoading] = useState(true);
@@ -27,53 +22,22 @@ export default function App() {
   const [searchQuery, setSearchQuery] = useState("");
   const [noResults, setNoResults] = useState(false);
 
-useEffect(() => {
-  setLoading(true);
-
-  const query = searchQuery || defaultQuery;
-  clientAPI.photos
-    .search({ query, per_page: PHOTOS })
-    .then(async response => { 
-
-      var myHeaders = new Headers();
-      myHeaders.append("x-api-key", "4d552e9f30522b1dec7c712f83c67c235be86e25ec14b4ba3493383ec7b81d3f");
-
-      let fetchPromises = await Promise.all(response.photos.map(async photo => { 
-        var formdata = new FormData();
-        formdata.append("image_url", photo.src.original);
-
-        var requestOptions = {
-          method: 'POST',
-          headers: myHeaders,
-          body: formdata,
-        };
-
-        try {
-          const fetchResponse = await fetch('https://cloudapi.lykdat.com/v1/detection/items', requestOptions);
-          const result = await fetchResponse.json();
-
-          if (result.data.detected_items.length !== 0) {
-            return photo;
-          }
-          
-        } catch (error) {
-          console.error('error', error);
-        }
-      }));
-
-      const fashionPhotos = fetchPromises.filter(photo => photo !== undefined);
-      setPhotos(fashionPhotos);
-      setNoResults(fashionPhotos.length === 0);
-      setLoading(false);
-    })
-    .catch(error => {
-      console.error("Error fetching photos:", error);
-      setLoading(false);
-    });
-
-  localStorage.setItem('user', JSON.stringify(user));
-}, [searchQuery]);
-
+  useEffect(() => {
+    setLoading(true);
+    fetch(`http://localhost:3000/photos?search=${searchQuery}`)
+      .then(response => response.json())
+      .then(photos => {
+        setLoading(false);
+        setPhotos(photos);
+        setNoResults(photos.length === 0);
+      })
+      .catch(error => {
+        console.error("Error fetching photos:", error);
+        setLoading(false);
+      });
+  
+    localStorage.setItem('user', JSON.stringify(user));
+  }, [searchQuery]);
 
   const handleSearch = query => {
     setSearchQuery(query);
